requirements.txt
reportportal-client
robotframework>=4
robotframework-debuglibrary
robotframework-seleniumlibrary
robotframework-jupyterlibrary>=0.3.1
ipython
openshift
pre-commit
pytest
pytest-logger
pyyaml
pygments
requests
Jinja2
flake8
mypy
kubernetes
validators
